pyproject.toml
[project]
name = "optimg"
version = "0.1.0"
description = "A simple image optimization tool"
readme = "README.md"
authors = [
    { name = "davidohnee", email = "dxstiny.gh@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "pillow>=12.0.0",
    "tqdm>=4.67.1",
]

[project.scripts]
optimg = "optimg:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[classifiers]
"Typing :: Typed" = "True"

[package]
include = ["py.typed"]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "ruff>=0.14.2",
]
